iota_sdk_crypto

Trait FromMnemonic

pub trait FromMnemonic {
    type Error;

    // Required methods
    fn from_mnemonic(
        phrase: &str,
        account_index: impl Into<Option<u64>>,
        password: impl Into<Option<String>>,
    ) -> Result<Self, Self::Error>
       where Self: Sized;
    fn from_mnemonic_with_path(
        phrase: &str,
        path: String,
        password: impl Into<Option<String>>,
    ) -> Result<Self, Self::Error>
       where Self: Sized;
}

Available on crate feature mnemonic only.

Defines a type which can be constructed from a mnemonic phrase

Required Associated Types

type Error

Required Methods

fn from_mnemonic( phrase: &str, account_index: impl Into<Option<u64>>, password: impl Into<Option<String>>, ) -> Result<Self, Self::Error>
where Self: Sized,

Create an instance from a mnemonic phrase

fn from_mnemonic_with_path( phrase: &str, path: String, password: impl Into<Option<String>>, ) -> Result<Self, Self::Error>
where Self: Sized,

Create an instance from a mnemonic phrase and a derivation path like:

Ed25519: "m/44'/4218'/0'/0'/0'"

Secp256k1: "m/54'/4218'/0'/0/0"

Secp256r1: "m/74'/4218'/0'/0/0"

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".