iota_sdk_crypto

Trait IotaVerifier

pub trait IotaVerifier {
    // Required methods
    fn verify_transaction(
        &self,
        transaction: &Transaction,
        signature: &UserSignature,
    ) -> Result<(), SignatureError>;
    fn verify_personal_message(
        &self,
        message: &PersonalMessage<'_>,
        signature: &UserSignature,
    ) -> Result<(), SignatureError>;
}

Interface for verifying user transactions and messages in IOTA

Note

This trait is implemented by the verifier and verifying key types of this crate as well as directly by the public key types of iota-sdk-types. The public key implementations validate the key bytes on every call, so when verifying many signatures with the same key it is cheaper to construct the corresponding verifying key once and verify with that instead.

Required Methods

fn verify_transaction( &self, transaction: &Transaction, signature: &UserSignature, ) -> Result<(), SignatureError>

fn verify_personal_message( &self, message: &PersonalMessage<'_>, signature: &UserSignature, ) -> Result<(), SignatureError>

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".