iota_sdk_crypto

Trait ToFromBase64

pub trait ToFromBase64 {
    type Error;

    // Required methods
    fn to_base64(&self) -> String;
    fn from_base64(value: &str) -> Result<Self, Self::Error>
       where Self: Sized;
}

Available on crate features ed25519 or secp256k1 or secp256r1 only.

Defines a type which can be converted to and from a base64 string of its raw bytes

Required Associated Types

type Error

Required Methods

fn to_base64(&self) -> String

Encode this private key as a base64 string of its raw bytes, without a signature scheme flag

fn from_base64(value: &str) -> Result<Self, Self::Error>
where Self: Sized,

Decode a private key from a base64 string of its raw bytes, without a signature scheme flag

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".