iota_sdk_crypto

Trait ToFromFlaggedBytes

pub trait ToFromFlaggedBytes {
    type Error;

    // Required methods
    fn to_flagged_bytes(&self) -> Vec<u8> ;
    fn from_flagged_bytes(bytes: &[u8]) -> Result<Self, Self::Error>
       where Self: Sized;
}

Defines a type that can be converted to and from flagged bytes, i.e. bytes prepended by some variant indicator flag

Required Associated Types

type Error

Required Methods

fn to_flagged_bytes(&self) -> Vec<u8>

Returns the bytes with the flag prepended

fn from_flagged_bytes(bytes: &[u8]) -> Result<Self, Self::Error>
where Self: Sized,

Creates an instance from bytes that include the flag