iota_sdk_crypto

Trait Verifier

pub trait Verifier<S> {
    // Required method
    fn verify(
        &self,
        message: &[u8],
        signature: &S,
    ) -> Result<(), SignatureError>;
}

Verify a signature of type S over a message.

Required Methods

fn verify(&self, message: &[u8], signature: &S) -> Result<(), SignatureError>

Verify that signature is a valid signature over message.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".